refactor(tokenizer): Extract NumberParser into own file
